Refactoring - Move listeners from CLI to logic.
